requirements.txt
pyfiglet==1.0.2
asyncio==3.4.3
pyrofork==2.3.61
tgcrypto-pyrofork==1.2.7
configparser==7.2.0
python-i18n==0.3.9
PyYAML==6.0.2
aiohttp==3.12.5
